Add ability to set autoexposure compensation in ros parameters
  - Add new `sensor::ae_compensation` parameter to set autoexposure brightness
  - Limit range from -9 to 9 (-3EV to +3EV)
  - Add ability to set dynamically while the node is running

[ROG-238]

+++ depthai_ros_driver/include/depthai_ros_driver/param_handlers/base_param_handler.hpp
@@ -4,7 +4,7 @@
 #include "rclcpp/node.hpp"
 namespace depthai_ros_driver {
 namespace param_handlers {
-inline rcl_interfaces::msg::ParameterDescriptor getRangedIntDescriptor(uint16_t min, uint16_t max) {
+inline rcl_interfaces::msg::ParameterDescriptor getRangedIntDescriptor(int32_t min, int32_t max) {
     {
         rcl_interfaces::msg::ParameterDescriptor desc;
         desc.integer_range.resize(1);
